requirements.txt
beautifulsoup4==4.12.3
diskcache==5.6.3
ipython==8.12.3
rich==13.7.0
selenium==4.9
undetected_chromedriver==3.5.4
watchdog==4.0.0
patool==1.12
openai==1.11.1
python-dotenv==1.0.1
pyyaml==6.0.1
ipython==8.12.3
tiktoken==2.31.0
autoflake==2.2.1
chardet==5.2.0